Generate random enemies
