+++ packages/client/src/pages/Game/Game.tsx
@@ -1,12 +1,17 @@
 import { FC, useEffect, useRef } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 import { Header } from '@components/Header';
 
+import { RoutePaths } from '@src/providers/Router/AppRouter/constants';
+
 import { CANVAS_HEIGHT, CANVAS_WIDTH } from './constants';
-import { handleKeyDown, requestAnimation } from './helpers';
+import { handleKeyDown } from './eventListeners';
+import { requestAnimation } from './mainEntry';
 
 export const Game: FC = () => {
   const canvasRef = useRef<HTMLCanvasElement>(null);
+  const navigate = useNavigate();
 
   useEffect(() => {
     const canvas = canvasRef.current;
@@ -19,9 +24,12 @@ export const Game: FC = () => {
   }, []);
 
   useEffect(() => {
-    window.addEventListener('keydown', handleKeyDown);
+    const listener = (e: KeyboardEvent) => {
+      handleKeyDown(e, () => navigate(RoutePaths.GAME_END));
+    };
+    window.addEventListener('keydown', listener);
     return () => {
-      window.removeEventListener('keydown', handleKeyDown);
+      window.removeEventListener('keydown', listener);
     };
   }, []);
