+++ packages/client/src/pages/Game/constants.ts
@@ -9,6 +9,7 @@ export const HOLE_RADIUS = CANVAS_WIDTH * 0.1;
 export const MIN_NUMBER_OF_ENEMIES = 5;
 export const MIN_RATIO_TO_HOLE = 0.4;
 export const MAX_RATIO_TO_HOLE = 2.25;
+export const HOLE_RADIUSES_PADDING = 4;
 
 export const MOVE_STEP = 10;
 export const GAME_ENTITY_FONT = '16px Comic Sans MS';
