+++ packages/client/src/pages/Game/gameState.ts
@@ -7,11 +7,5 @@ export const gameState: TGameState = {
     x: 100,
     y: 100,
   },
-  enemies: [
-    { x: 150, y: 200, points: 10, radius: 10, isVisible: true },
-    { x: 200, y: 200, points: 20, radius: 20, isVisible: true },
-    { x: 100, y: 250, points: 30, radius: 30, isVisible: true },
-    { x: 500, y: 400, points: 125, radius: 125, isVisible: true },
-    { x: 600, y: 200, points: 30, radius: 30, isVisible: true },
-  ],
+  enemies: [],
 };
